MediTrack/requirements.txt
pandas==2.2.2
numpy==1.26.4
Flask==2.3.3
PyMySQL==1.1.0
bcrypt==4.0.1
Jinja2==3.1.2
Werkzeug==2.3.7
click==8.1.7
python-dotenv==1.0.0
gunicorn==21.2.0
pdfplumber==0.9.0
PyPDF2==3.0.1
python-docx==0.8.11

Pillow==10.4.0
pytesseract==0.3.13
easyocr==1.7.1
scipy==1.11.4
paddleocr==2.7.0.3
transformers==4.35.0
torch==2.3.1
